+++ libcsight/include/csight/PKM.hpp
@@ -19,7 +19,7 @@ namespace csight {
     virtual std::string GetAbilityString() = 0;
     virtual u32 GetPID() = 0;
     virtual u64 GetRaidSeed() = 0;
-    virtual u8 GetForm() = 0;
+    virtual u16 GetForm() = 0;
     u32 GetPSV() {
       u32 pid = this->GetPID();
       return ((pid >> 16 ^ (pid & 0xFFFF)) >> 4);
